feat: interfaces.goをrepository層にも追加
